+++ examples/lang/js_web_api/build-web.sh
@@ -12,6 +12,7 @@ WEB_DIR="./web/"
 WASM="${CRATE_NAME}.wasm"
 TARGET_DIR="target/wasm32-unknown-unknown/${PROFILE}/"
 JS_DIR="../../../src/lang/js/"
+# JS_URL="https://raw.githubusercontent.com/andamira/devela/refs/heads/main/src/lang/js/web_api.js"
 
 CARGO_TARGET_DIR="./target/" cargo b --profile $PROFILE --target wasm32-unknown-unknown
 
@@ -24,6 +25,7 @@ cp "${TARGET_DIR}/${WASM}" $WEB_DIR
 # js library
 cp "${JS_DIR}web_api.js" $WEB_DIR
 # uglifyjs "${JS_DIR}web_api.js" -o "${WEB_DIR}web_api.js"
+# curl -s "${JS_URL}" > "${WEB_DIR}web_api.js"
 
 chmod -x "${WEB_DIR}${WASM}"